changes to file structure
